[refactoring] Ladder 객체의 메소드 리펙토링, 분리
